调整 User 项目结构 增加管理后台查看 User 分页接口

+++ user/user-service-api/src/main/java/cn/iocoder/mall/user/service/api/bo/OAuth2AuthenticationBO.java
@@ -7,13 +7,13 @@ public class OAuth2AuthenticationBO implements Serializable {
     /**
      * 用户编号
      */
-    private Long uid;
+    private Integer uid;
 
-    public Long getUid() {
+    public Integer getUid() {
         return uid;
     }
 
-    public OAuth2AuthenticationBO setUid(Long uid) {
+    public OAuth2AuthenticationBO setUid(Integer uid) {
         this.uid = uid;
         return this;
     }
